+++ tests/trainer/test_freq_accumulation.py
